requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2
pymysql
pandas
scikit-learn
joblib
python-dotenv
